finishing and debugging TauP travel time calculator
