+++ src/main/java/globalquake/main/Main.java
@@ -5,6 +5,7 @@ import globalquake.database.StationDatabaseManager;
 import globalquake.database.StationSource;
 import globalquake.exception.ApplicationErrorHandler;
 import globalquake.exception.FatalIOException;
+import globalquake.geo.taup.TauPTravelTimeCalculator;
 import globalquake.regions.Regions;
 import globalquake.sounds.Sounds;
 import globalquake.ui.database.DatabaseMonitorFrame;
@@ -63,18 +64,21 @@ public class Main {
         databaseMonitorFrame.getMainProgressBar().setValue(0);
         Regions.init();
         databaseMonitorFrame.getMainProgressBar().setString("Loading scales...");
-        databaseMonitorFrame.getMainProgressBar().setValue(20);
+        databaseMonitorFrame.getMainProgressBar().setValue((int) (1 / 6.0 * 100.0));
         Scale.load();
         databaseMonitorFrame.getMainProgressBar().setString("Loading sounds...");
-        databaseMonitorFrame.getMainProgressBar().setValue(40);
+        databaseMonitorFrame.getMainProgressBar().setValue((int) (2 / 6.0 * 100.0));
         Sounds.load();
+        databaseMonitorFrame.getMainProgressBar().setString("Loading travel table...");
+        databaseMonitorFrame.getMainProgressBar().setValue((int) (3 / 6.0 * 100.0));
+        TauPTravelTimeCalculator.init();
         databaseMonitorFrame.getMainProgressBar().setString("Updating Station Sources...");
-        databaseMonitorFrame.getMainProgressBar().setValue(60);
+        databaseMonitorFrame.getMainProgressBar().setValue((int) (4 / 6.0 * 100.0));
         databaseManager.runUpdate(databaseManager.getStationDatabase().getStationSources().stream()
                         .filter(StationSource::isOutdated).collect(Collectors.toList()),
                 () -> {
                     databaseMonitorFrame.getMainProgressBar().setString("Checking Seedlink Networks...");
-                    databaseMonitorFrame.getMainProgressBar().setValue(80);
+                    databaseMonitorFrame.getMainProgressBar().setValue((int) (5 / 6.0 * 100.0));
                     databaseManager.runAvailabilityCheck(databaseManager.getStationDatabase().getSeedlinkNetworks(), () -> {
                         databaseMonitorFrame.getMainProgressBar().setString("Done");
                         databaseMonitorFrame.getMainProgressBar().setValue(100);
